Moves essdive_trasfer package under service
Issue #372
